moved list into new TodoList Component, created AddTodoForm Component, and bonus-- added new Component called MotivationalQuote to App.jsx
