+++ src/MotivationalQuote.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function MotivationalQuote () {
+    return (
+        <h3><em> "First, forget inspiration. Habit is more dependable. Habit will sustain you
+        whether you're inspired or not. Habit will help you finish and polish your stories. 
+        Inspiration won't. Habit is persistence in practice." <br></br> - Octavia Butler
+        </em></h3>
+    )
+}
+
+
+
+export default MotivationalQuote;
